GLBBuilder: Small API adjustments
